move tests to tx_progress file
